load_diploma_servise/requirements.txt
demjson3==3.0.6
python-dotenv==1.0.1
redis==5.2.1
Requests==2.32.3
